pyconby-web-api-from-scratch-with-fastapi/apiapp/requirements.txt
fastapi
uvicorn
sqlalchemy
async-exit-stack
async-generator
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
